% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Stay/Switch door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{arguments used: first if the contestant chooses to stay their final selection will
equal to their initial selection, second if the contestant chooses to change to the
remaining other door their final pick will be different than their initial selection.}
}
\value{
The function returns the game, initial door number picked, final door number picked.
}
\description{
\code{change_door()} contestant chooses to switch his/her initial selection or stay.
}
\details{
The contestant chooses to stay on his/her initial selection, or to switch to the other
door that is different from the opened one and from their initial pick.
}
\examples{
  opened.door <- open_goat_door( this.game, my.initial.pick )
  change_door( stay=T, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  change_door( stay=F, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  my.final.pick <- change_door( stay=F, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  this.game
  my.initial.pick
  my.final.pick
}
