% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determining the winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{arguments used: first if the final pick turns out a car the contestant will win, and if the
final pick turns out to be a goat the contestant will loose.}
}
\value{
The function returns the game, initial door number picked, final door number picked.
}
\description{
\code{determine_winner()} see if the contestant has won or lost.
}
\details{
Check the final pick of the contestant if it's a car he/she wins, otherwise he/she loses.
}
\examples{
  my.initial.pick
  my.final.pick <- change_door( stay=T, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  determine_winner( final.pick=my.final.pick, 
  game=this.game )
  my.final.pick <- change_door( stay=F, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  determine_winner( final.pick=my.final.pick, 
  game=this.game )

}
