% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{2 arguments used: first if the contestant selected car above then the host can choose
any of the 2 goats, and second if the contestant selected car above.}
}
\value{
The function returns the game, initial door number picked, opened door number by
the host(goat).
}
\description{
\code{open_goat_door()} picks a door with a goat behind.
}
\details{
The Host chooses a door with a goat behind that is different from the door picked
by the contestant and of course not having a car behind.
}
\examples{
  this.game <- create_game()
  this.game
  my.initial.pick <- select_door()
  my.initial.pick
  open_goat_door( this.game, my.initial.pick )
}
