% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Mounty Hall Game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the game results if the contestant switched what would
be the result (win or lose) and if contestant stayed what would
be the result (win or lose)
}
\description{
\code{play_game()} is a function including all game steps together.
}
\details{
It's a wrapper function that makes it easier to control and repeat the steps of the game
in your simulation
}
\examples{
  play_game()

}
