% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a random door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector
indicating the door that the contestant choose.
}
\description{
\code{select_door()} picks a random door that have either a goat or car behind.
}
\details{
The contestant selects a door (picked randomly here) and he/she cannot see what's behind.
}
\examples{
  select_door()

}
