% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Contestant given the opportunity to switch their initial door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{function uses the strategy used as an argument stay = TRUE
or stay = FALSE. It takes into account the door revealed by the
host, opened.door, and the initial door, a.pick.}
}
\value{
The function returns a number between 1 and 3.
}
\description{
change_door is a function that shows the contestant's strategy of
switching or keeping their door.
}
\details{
After a goat door is revealed only the door initially selected and the other
closed door remain. The contestant will have the opportunity to keep their
initial door or change their door.
}
\examples{
change_door(stay = TRUE, opened.door, a.pick)
}
