% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether contestant is a winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{function takes into account the final.pick and the game.}
}
\value{
The function returns "WIN" or "LOSE"
}
\description{
determine_winner is a function that reveal whether the contestant's
door is winner; a car behind it.
}
\details{
The contestant's door is revealed and it will be known whether the door
had a goat or a car behind it.
}
\examples{
determine_winner(final.pick, game)
}
