% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host reveals a goat door to the contestant.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function uses the return value, a.pick, from select_door function}
}
\value{
The function returns a number between 1 and 3.
}
\description{
open_goat_door is a function that reveals a door that contains a goat.
}
\details{
The host will open a door with a goat. If the initial door selected
is a car either remaining goat will be opened at random. If a goat
door is initially selected the door the host will reveal the remaining
goat door.
}
\examples{
open_goat_door(game, a.pick)
}
