% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Game results}
\usage{
play_game()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a data frame showing the outcome by strategy.
}
\description{
The function play_game plays the created game and gives you the
outcome of the game.
}
\details{
The function runs the game and gives the strategy and outcome in
a data frame.
}
\examples{
play_game()
}
