% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Looping the game.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{function takes the number of games to be run. 100 is default.}
}
\value{
The function returns a data frame with average win rate by outcome and
the results of each game by strategy.
}
\description{
play_n_games() runs the game n number of times and gives you outcome
outcome by strategy.
}
\details{
The game is simulated will allow users to run the game n number of times.
This will allows is to test the average win rate by strategy and let us
determine whcih strategy is dominant.
}
\examples{
play_n_games(n=10000)
}
