% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3.
}
\description{
select_door is a function that select a door at random.
}
\details{
The game consists of three doors and the contestant selects
their initial door.
}
\examples{
select_door()
}
